feat(cli): expose AccountPacket APIs from a lib so it can be used by other apps
