+++ sn_cli/Cargo.toml
@@ -14,6 +14,10 @@ version = "0.89.85"
 path="src/main.rs"
 name="safe"
 
+[lib]
+path="src/lib/lib.rs"
+name="autonomi"
+
 [[bench]]
 name = "files"
 harness = false
