+++ sn_cli/src/lib/lib.rs
@@ -1,0 +1,16 @@
+// Copyright 2024 MaidSafe.net limited.
+//
+// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
+// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
+// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied. Please review the Licences for the specific language governing
+// permissions and limitations relating to use of the SAFE Network Software.
+
+mod acc_packet;
+mod files;
+
+pub use acc_packet::AccountPacket;
+pub use files::{
+    chunks_to_upload_with_iter, download_file, download_files, ChunkManager, Estimator,
+    FilesUploadOptions, IterativeUploader, UploadedFile, UPLOADED_FILES,
+};
