+++ sn_cli/src/subcommands.rs
@@ -6,7 +6,6 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-mod acc_packet;
 pub(crate) mod files;
 pub(crate) mod folders;
 pub(crate) mod register;
