+++ sn_cli/src/subcommands/folders.rs
@@ -6,13 +6,13 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::subcommands::acc_packet::AccountPacket;
+use autonomi::{AccountPacket, FilesUploadOptions};
 
-use sn_client::{Client, BATCH_SIZE};
+use sn_client::{
+    protocol::storage::{RegisterAddress, RetryStrategy},
+    Client, BATCH_SIZE,
+};
 
-use sn_client::protocol::storage::{RegisterAddress, RetryStrategy};
-
-use crate::subcommands::files::upload::FilesUploadOptions;
 use clap::Parser;
 use color_eyre::{eyre::eyre, Result};
 use std::{
